api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
pydantic==2.9.2
python-dotenv==1.0.1
flywheel-sdk==20.6.0
httpx==0.27.2
pytest==8.3.3
pillow
nibabel

